clean test_app setup task, issue with db test setup

+++ core/lib/gaku/testing/common_rake.rb
@@ -21,53 +21,53 @@ namespace :common do
 
     puts 'Setting up dummy database...'
 
-    db_accessible = false
-    begin
-      ActiveRecord::Base.establish_connection
-      ActiveRecord::Base.connection
-      db_accessible = true if ActiveRecord::Base.connected?
-    rescue
+    if (!!(`PGPASSWORD=manabu psql postgres --user manabu -c "SELECT current_setting('is_superuser');"` =~ /on/))
+      `bundle exec rails app:update:bin db:environment:set db:drop db:create db:migrate db:test:prepare RAILS_ENV=test`
+    else
       printf "Database was not accessible. How would you like to proceed?\n" \
-        "(1) Try to create user and database manually (requires sudo)\n" \
-        "(2) Attempt to run migrations. This requires the following:\n" \
+        "(1) Try to create user and database automatically (requires sudo)\n" \
+        "(2) Attempt to run creation tasks/migrations. This requires the following:\n" \
         "\t * HStore installed\n" \
         "\t * Postgres user \"manabu\" with password \"manabu\"\n" \
         "\t * Database \"gaku_test\" and \"gaku_development\" with full permission granted to \"manabu\"\n" \
-        "(3) Use \"manabu\" Postgres user (password \"manabu\"), assuming user has superuser status\n" \
-        "Please enter 1, 2, or 3. Enter anything else or simply hit enter to cancel: "
+        "Please enter 1 or 2. Enter anything else or simply hit enter to cancel: "
       selection = STDIN.getc
-      puts "Selection was #{selection.chr}"
-
-      `sudo -u postgres psql -c "CREATE USER manabu WITH PASSWORD 'manabu';"`
-      `sudo -u postgres psql -c "ALTER USER manabu CREATEDB;"`
-      `sudo -u postgres psql -c "CREATE EXTENSION IF NOT EXISTS hstore;"`
-      
-      #`sudo -u postgres psql -c "DROP DATABASE gaku_test;"`
-      #`sudo -u postgres psql -c "CREATE DATABASE gaku_test;"`
-      #`sudo -u postgres psql -c "GRANT ALL PRIVILEGES ON DATABASE \"gaku_test\" to manabu;"`
-      #`sudo -u postgres psql -c "ALTER DATABASE gaku_test OWNER TO manabu;"`
-      #`sudo -u postgres psql -c "DROP DATABASE gaku_development;"`
-      #`sudo -u postgres psql -c "CREATE DATABASE gaku_development;"`
-      #`sudo -u postgres psql -c "GRANT ALL PRIVILEGES ON DATABASE \"gaku_development\" to manabu;"`
-      #`sudo -u postgres psql -c "ALTER DATABASE gaku_development OWNER TO manabu;"`
-      
-      #`sudo -u postgres psql -c "ALTER ROLE manabu superuser;"`
-      `bundle exec rails app:update:bin db:environment:set db:drop db:create db:migrate db:test:prepare RAILS_ENV=test`
-      #`sudo -u postgres psql -c "ALTER ROLE manabu nosuperuser;"`
-    end
-
-    if db_accessible
-      puts 'Database found and is accessible by manabu user. Assuming empty and preparing. To clear database run "rake drop_test_app".'
-      #cmd = 'bundle exec rails app:update:bin db:environment:set db:migrate db:test:prepare RAILS_ENV=test'
-      #cmd = 'bundle exec rails app:update:bin db:environment:set db:drop db:create db:migrate db:test:prepare RAILS_ENV=test'
-
-      #if RUBY_PLATFORM =~ /mswin/ # windows
-      #  cmd += ' >nul'
-      #else
-      #  cmd += ' >/dev/null'
-      #end
-
-      #system(cmd)
+      case selection
+        when '1' then
+          puts 'Creating user and database automatically...'
+          _autopilot_setup
+        when '2' then
+          puts 'Running creation tasks/migrations...'
+          `bundle exec rails app:update:bin db:environment:set db:migrate db:test:prepare RAILS_ENV=test`
+        else
+          puts
+      end
     end
   end
+end
+
+def _autopilot_setup
+  term_cover = " 1> /dev/null 2> /dev/null"
+  puts "Setting up DB User, Database, and Extensions..."
+  `sudo -u postgres psql -c "CREATE USER manabu WITH PASSWORD 'manabu';"#{term_cover}`
+  `sudo -u postgres psql -c "ALTER USER manabu CREATEDB;"#{term_cover}`
+  `sudo -u postgres psql -c "CREATE EXTENSION IF NOT EXISTS hstore;"#{term_cover}`
+
+  `sudo -u postgres psql -c "DROP DATABASE gaku_test;"#{term_cover}`
+  `sudo -u postgres psql -c "CREATE DATABASE gaku_test;"#{term_cover}`
+  `sudo -u postgres psql -c "GRANT ALL PRIVILEGES ON DATABASE \"gaku_test\" to manabu;"#{term_cover}`
+  `sudo -u postgres psql gaku_test -c "CREATE EXTENSION hstore;"#{term_cover}`
+  `sudo -u postgres psql -c "ALTER DATABASE gaku_test OWNER TO manabu;"#{term_cover}`
+  
+  `sudo -u postgres psql -c "DROP DATABASE gaku_development;"#{term_cover}`
+  `sudo -u postgres psql -c "CREATE DATABASE gaku_development;"#{term_cover}`
+  `sudo -u postgres psql -c "GRANT ALL PRIVILEGES ON DATABASE \"gaku_development\" to manabu;"#{term_cover}`
+  `sudo -u postgres psql gaku_development -c "CREATE EXTENSION hstore;"#{term_cover}`
+  `sudo -u postgres psql -c "ALTER DATABASE gaku_development OWNER TO manabu;"#{term_cover}`
+  puts "Done."
+
+  puts "Running tasks..."
+  `bundle exec rails app:update:bin db:environment:set db:migrate db:test:prepare RAILS_ENV=test`
+  puts "Done. If the test app does not run normally please follow the setup guide at: "
+  puts "https://github.com/GAKUEngine/gaku"
 end
